# Conflicts:
#	src/main/resources/static/js/change/changeForm/changeOrder.js
#	src/main/resources/templates/bomManage/ebom/ebomManage/ebomSetChangeForm.html
#	src/main/resources/templates/change/changeOrder/changeOrder.html

+++ src/main/java/sql/pojo/cfg/fullCfg/HzFullCfgWithCfg.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2018.
- * This file was wrote by fancyears·milos·malvis @connor. Any question/bug you can post to 1243093366@qq.com.
+ * This file was written by fancyears·milos·malvis @connor. Any question/bug you can post to 1243093366@qq.com.
  * ALL RIGHTS RESERVED.
  */
 
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 import java.util.Date;
 
 /**
- * @Author: Fancyears·Maylos·Maywas
+ * @Author: Fancyears·Maylos·Malvis
  * @Description: 全配置BOM一级清单2Y与特性值关联
  * @Date: Created in 2018/8/10 13:55
  * @Modified By:
@@ -32,7 +32,7 @@ public class HzFullCfgWithCfg {
 
     private String flCfgBomlineName;
 
-    private BigDecimal flCfgVersion;
+    private Long flCfgVersion;
     /**
      * 操作类型,1新增，2更新，3删除
      */
@@ -106,11 +106,11 @@ public class HzFullCfgWithCfg {
         this.flCfgBomlineName = flCfgBomlineName;
     }
 
-    public BigDecimal getFlCfgVersion() {
+    public Long getFlCfgVersion() {
         return flCfgVersion;
     }
 
-    public void setFlCfgVersion(BigDecimal flCfgVersion) {
+    public void setFlCfgVersion(Long flCfgVersion) {
         this.flCfgVersion = flCfgVersion;
     }
