+++ src/main/java/sql/pojo/cfg/modelColor/HzCmcrDetailChange.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2018.
- * This file was wrote by fancyears·milos·malvis @connor. Any question/bug you can post to 1243093366@qq.com.
+ * This file was written by fancyears·milos·malvis @connor. Any question/bug you can post to 1243093366@qq.com.
  * ALL RIGHTS RESERVED.
  */
 
@@ -11,7 +11,7 @@ import lombok.Data;
 import java.math.BigDecimal;
 import java.util.Date;
 /**
- * @Author: Fancyears·Maylos·Maywas
+ * @Author: Fancyears·Maylos·Malvis
  * @Description: 配色方案变更详情数据
  * @Date: Created in 2018/8/10 13:55
  * @Modified By:
@@ -133,5 +133,5 @@ public class HzCmcrDetailChange {
     /**
      *变更主表ID
      */
-    private Integer cmcrCgChangeId;
+    private Long cmcrCgChangeId;
 }
